Revert "Pull request of new interactive attention head visualization feature"
